+++ src/schemes/schemes.jl
@@ -9,10 +9,13 @@ include("structure/discrete_element_method/discrete_element_method.jl")
 # Monaghan-Kajtar repulsive boundary particles require the `WallBoundarySystem`
 # and the `TotalLagrangianSPHSystem`.
 include("boundary/wall_boundary/monaghan_kajtar.jl")
+# Implicit incompressible SPH requires the `BoundarySPHSystem`
+include("fluid/implicit_incompressible_sph/implicit_incompressible_sph.jl")
 
 # Include rhs for all schemes
 include("fluid/weakly_compressible_sph/rhs.jl")
 include("fluid/entropically_damped_sph/rhs.jl")
+include("fluid/implicit_incompressible_sph/rhs.jl")
 include("boundary/wall_boundary/rhs.jl")
 include("structure/total_lagrangian_sph/rhs.jl")
 include("structure/discrete_element_method/rhs.jl")