+++ NEWS.md
@@ -4,6 +4,14 @@ TrixiParticles.jl follows the interpretation of
 [semantic versioning (semver)](https://julialang.github.io/Pkg.jl/dev/compatibility/#Version-specifier-format-1)
 used in the Julia ecosystem. Notable changes will be documented in this file for human readability.
 
+## Version 0.4.1
+
+### Features
+
+- Added an incompressible SPH solver `ImplicitIncompressibleSPHSystem` (#751).
+
+
+
 ## Version 0.4
 
 ### API Changes